+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add QuantityKinds and Units required to map IFC to QUDT. These are added temporarily until an upstream release of QUDT.
 
+### Changed
+
+- Modify the selection of units matching a set of factor units. The matching algorithm as well as the scoring algorithm of matching units was adapted. With this change, there are only two options for the client: get all matching units or get the one that matches best.
+
 ## [2.2.0] - 2022-10-28
 
 ### Changed